# contracts/salesagent_api_contract.yaml
openapi: 3.1.0
info:
  title: Sales Agent E-Commerce Sales Concierge API
  version: 1.0.0
  description: V1 application and evaluation testability contract. The chat endpoint is customer-facing; the trace endpoint is intended for local/evaluation environments only.
paths:
  /api/v1/chat:
    post:
      summary: Send a shopper message to the concierge
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              new_session:
                value:
                  message: I need a waterproof jacket for Scotland under £160
              existing_session:
                value:
                  session_id: session-123
                  message: I'd prefer Ocean Blue in size M
      responses:
        '200':
          description: Successful concierge response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
        '422':
          description: Schema validation failure
        '500':
          description: Unexpected server error
  /api/v1/traces/{trace_id}:
    get:
      summary: Retrieve an evaluation trace for a completed chat turn
      description: Evaluation/local-only endpoint. Must be disabled or protected in a production-style deployment.
      operationId: getTrace
      parameters:
      - name: trace_id
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Structured execution trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceResponse'
        '404':
          description: Trace not found
components:
  schemas:
    ChatRequest:
      type: object
      additionalProperties: false
      required:
      - message
      properties:
        session_id:
          type:
          - string
          - 'null'
          minLength: 1
          description: Omit or send null to create a new session.
        message:
          type: string
          minLength: 1
          maxLength: 4000
    ChatResponse:
      type: object
      additionalProperties: false
      required:
      - session_id
      - trace_id
      - message
      - recommendations
      properties:
        session_id:
          type: string
        trace_id:
          type: string
        message:
          type: string
          description: Natural-language shopper-facing response.
        recommendations:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/ProductRecommendation'
        promotion:
          anyOf:
          - $ref: '#/components/schemas/PromotionResult'
          - type: 'null'
          default: null
        pricing:
          anyOf:
          - $ref: '#/components/schemas/PricingResult'
          - type: 'null'
          default: null
    ProductRecommendation:
      type: object
      additionalProperties: false
      required:
      - product_id
      - name
      - price
      - currency
      - product_url
      - availability
      properties:
        product_id:
          type: string
        name:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
          const: GBP
        product_url:
          type: string
        availability:
          type: string
          enum:
          - in_stock
          - out_of_stock
          - partial
          - unknown
        matched_variant:
          anyOf:
          - $ref: '#/components/schemas/VariantAvailability'
          - type: 'null'
          default: null
    PromotionResult:
      type: object
      additionalProperties: false
      required:
      - code
      - valid
      properties:
        code:
          type: string
        valid:
          type: boolean
        discount_percent:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 100
        reason:
          type:
          - string
          - 'null'
          description: 'Examples: active, inactive, unknown_code.'
    PricingResult:
      type: object
      additionalProperties: false
      required:
      - product_id
      - base_price
      - final_price
      - currency
      properties:
        product_id:
          type: string
        base_price:
          type: number
          minimum: 0
        final_price:
          type: number
          minimum: 0
        currency:
          type: string
          const: GBP
        discount_code:
          type:
          - string
          - 'null'
        discount_percent:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 100
    TraceResponse:
      type: object
      additionalProperties: false
      required:
      - trace_id
      - session_id
      - timestamp
      - turn_index
      - user_message
      - model
      - prompt_version
      - resolved_constraints
      - constraint_changes
      - tool_calls
      - recommendation_validation
      - recommended_product_ids
      - latency_ms
      - token_usage
      - errors
      properties:
        trace_id:
          type: string
        session_id:
          type: string
        timestamp:
          type: string
          format: date-time
        turn_index:
          type: integer
          minimum: 1
        user_message:
          type: string
        model:
          type: string
        prompt_version:
          type: string
        resolved_constraints:
          $ref: '#/components/schemas/ResolvedConstraints'
        constraint_changes:
          type: array
          items:
            $ref: '#/components/schemas/ConstraintChange'
        tool_calls:
          type: array
          description: Ordered list of tool calls for this turn. Sequence is authoritative for trajectory evaluation.
          items:
            $ref: '#/components/schemas/ToolCallTrace'
        recommendation_validation:
          $ref: '#/components/schemas/RecommendationValidation'
        recommended_product_ids:
          type: array
          items:
            type: string
        promotion:
          anyOf:
          - $ref: '#/components/schemas/PromotionResult'
          - type: 'null'
          default: null
        pricing:
          anyOf:
          - $ref: '#/components/schemas/PricingResult'
          - type: 'null'
          default: null
        latency_ms:
          type: integer
          minimum: 0
        token_usage:
          $ref: '#/components/schemas/TokenUsage'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/TraceError'
    ResolvedConstraints:
      type: object
      additionalProperties: false
      properties:
        category:
          type:
          - string
          - 'null'
        activity:
          type:
          - string
          - 'null'
        weather:
          type: array
          items:
            type: string
        features:
          type: array
          items:
            type: string
        maximum_price:
          type:
          - number
          - 'null'
          minimum: 0
        colour:
          type:
          - string
          - 'null'
        size:
          type:
          - string
          - 'null'
        season:
          type:
          - string
          - 'null'
        priority:
          type:
          - string
          - 'null'
    ConstraintChange:
      type: object
      additionalProperties: false
      required:
      - field
      - previous
      - current
      properties:
        field:
          type: string
        previous: {}
        current: {}
    ToolCallTrace:
      type: object
      additionalProperties: false
      required:
      - sequence
      - tool_call_id
      - tool_name
      - arguments
      - result
      - status
      - duration_ms
      properties:
        sequence:
          type: integer
          minimum: 1
        tool_call_id:
          type: string
        tool_name:
          type: string
          enum:
          - search_products
          - get_product
          - check_inventory
          - validate_discount
        arguments:
          type: object
          description: Schema-validated arguments supplied to the commerce tool.
        result:
          type: object
          description: Authoritative structured commerce-tool result.
        status:
          type: string
          enum:
          - success
          - error
        duration_ms:
          type: integer
          minimum: 0
    RecommendationValidation:
      type: object
      additionalProperties: false
      required:
      - all_products_exist
      - prices_match_catalogue
      - urls_match_catalogue
      - stock_claims_validated
      properties:
        all_products_exist:
          type: boolean
        prices_match_catalogue:
          type: boolean
        urls_match_catalogue:
          type: boolean
        stock_claims_validated:
          type: boolean
        validation_errors:
          type: array
          items:
            type: string
          default: []
    VariantAvailability:
      type: object
      additionalProperties: false
      required:
      - colour
      - size
      - stock
      properties:
        colour:
          type: string
        size:
          type: string
        stock:
          type: integer
          minimum: 0
    TokenUsage:
      type: object
      additionalProperties: false
      required:
      - input_tokens
      - output_tokens
      - total_tokens
      properties:
        input_tokens:
          type: integer
          minimum: 0
        output_tokens:
          type: integer
          minimum: 0
        total_tokens:
          type: integer
          minimum: 0
    TraceError:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      properties:
        code:
          type: string
        message:
          type: string
        tool_call_id:
          type:
          - string
          - 'null'